serviceapi/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

google-cloud-storage>=2.14.0
google-auth>=2.0.0
firebase-admin>=6.0.0

gunicorn==21.2.0
cloud-sql-python-connector~=1.0
pg8000==1.30.3
SQLAlchemy==2.0.25
dotenv==1.2.2